Se agrega formulario para crear usuarios

+++ app/Http/Controllers/UsersController.php
@@ -2,7 +2,9 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Support\Facades\Auth;
 use App\User;
+use App\Language;
 
 class UsersController extends Controller
 {
@@ -12,11 +14,26 @@ class UsersController extends Controller
 
         return view('users.index', ['users' => $users]);
     }
-
-    public function destroy(User $user)
+    public function create(){
+        return view('users.nuevo');
+    }
+    public function store(){
+        //$this->authorize('create', User::class);
+        $admin = 0;
+        if(request('chkAdmin')=='on') $admin = 1;
+        User::create([
+            'name'              => request('name'),
+            'email'             => request('email'),
+            'locale'            => Language::ES,
+            'password'          => bcrypt(request('pass')),
+            'admin'             => $admin,
+        ]);
+        return redirect()->route('users.index');
+    }
+    public function delete(User $user)
     {
+        if(Auth::user()->id==$user->id) return back();
         $user->delete();
-
         return back();
     }
